+++ lib/td_dd_web/schema.ex
@@ -14,19 +14,23 @@ defmodule TdDdWeb.Schema do
   import_types(TdDdWeb.Schema.Sources)
   import_types(TdDdWeb.Schema.StructureNotes)
   import_types(TdDdWeb.Schema.Structures)
+  import_types(TdDdWeb.Schema.StructureTags)
   import_types(TdDdWeb.Schema.Templates)
   import_types(TdDdWeb.Schema.Types.Custom.JSON)
 
   query do
+    import_fields(:domain_queries)
     import_fields(:rule_queries)
     import_fields(:source_queries)
     import_fields(:structure_note_queries)
     import_fields(:structure_queries)
+    import_fields(:structure_tag_queries)
     import_fields(:template_queries)
   end
 
   mutation do
     import_fields(:source_mutations)
+    import_fields(:structure_tag_mutations)
   end
 
   def context(ctx) do